Add files via upload
These files are part of the integration of a Streamlit-based web user interface for hands-on interaction with the life expectancy prediction project. The package includes all the necessary components to run the project and visualize the machine learning models' outputs.

Additionally, a video tutorial (in MP4 format) is included, demonstrating the complete process of how to effectively use the files. The video walks through each step, from setting up the environment to running the Streamlit app, and explains how to input data, view predictions, and interact with the visualizations.

With this integration, users can seamlessly engage with the project, input custom data, and explore the results, providing an interactive way to understand the underlying models and data insights.